pyproject.toml
[project]
name = "financial-risk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.4.0",
    "faiss-cpu>=1.13.2",
    "groq>=1.0.0",
    "kafka-python>=2.3.0",
    "numpy>=2.2.6",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
